# docs/swagger.json
{
 "swagger": "2.0",
 "info": {
  "description": "Resource for flowline",
  "title": "Flowline REST API",
  "contact": {
   "name": "sysatom",
   "url": "https://flowline.tsundata.com",
   "email": "sysatom@gmail.com"
  },
  "license": {
   "name": "MIT",
   "url": "https://github.com/tsundata/flowline/blob/main/LICENSE"
  },
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/apps/v1/user/session": {
   "post": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "tags": [
     "user"
    ],
    "summary": "Create user resource to session subresource",
    "operationId": "userCreateSession",
    "parameters": [
     {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/meta.UserSession"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/meta.UserSession"
      }
     }
    }
   }
  },
...
 },
 "definitions": {
...
  "meta.UserSession": {
   "properties": {
    "apiVersion": {
     "type": "string"
    },
    "creationTimestamp": {
     "type": "string",
     "format": "date-time"
    },
    "deletionGracePeriodSeconds": {
     "type": "integer",
     "format": "int64"
    },
    "deletionTimestamp": {
     "type": "string",
     "format": "date-time"
    },
    "finalizers": {
     "type": "array",
     "items": {
      "type": "string"
     }
    },
    "generation": {
     "type": "integer",
     "format": "int64"
    },
    "kind": {
     "type": "string"
    },
    "labels": {
     "type": "object",
     "additionalProperties": {
      "type": "string"
     }
    },
    "name": {
     "type": "string"
    },
    "password": {
     "type": "string"
    },
    "resourceVersion": {
     "type": "string"
    },
    "token": {
     "type": "string"
    },
    "uid": {
     "type": "string"
    },
    "userUID": {
     "type": "string"
    },
    "username": {
     "type": "string"
    }
   }
  },
...
 },
 "securityDefinitions": {
  "BearerToken": {
   "description": "Bearer Token authentication",
   "type": "apiKey",
   "name": "authorization",
   "in": "header"
  }
 },
 "security": [
  {
   "BearerToken": []
  }
 ],
...
}